continues to test for bug, think solution found
I believe that the problem is caused by the mesh for the basis functions and the quadrature points not being aligned. This is causing some integrals to over estimate and some to underestimate (I believe). It seems a little odd that it is only underestimating in 2 places but that could be a problem and this way we will be giving the best solution as the integration will be supported exactly on the grid we need.